# openapi.with-responses.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/flashcards": {
      "post": {
        "operationId": "FlashcardsController_createFlashcard",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFlashcardDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/FlashcardResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "400 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "401 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "404 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "500 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Flashcards"
        ]
      },
      "get": {
        "operationId": "FlashcardsController_getFlashcards",
        "parameters": [
          {
            "name": "bookId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/FlashcardResponse"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "400 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "401 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "500 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Flashcards"
        ]
      }
    },
    "/api/v1/flashcards/review/due": {
      "get": {
        "operationId": "FlashcardsController_getCardsForReview",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/FlashcardResponse"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "401 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "500 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Flashcards"
        ]
      }
    },
    "/api/v1/flashcards/stats": {
      "get": {
        "operationId": "FlashcardsController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/FlashcardStatsResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "401 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "500 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Flashcards"
        ]
      }
    },
    "/api/v1/flashcards/{id}": {
      "get": {
        "operationId": "FlashcardsController_getFlashcardById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/FlashcardResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "401 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "403 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "404 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "500 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Flashcards"
        ]
      },
      "put": {
        "operationId": "FlashcardsController_updateFlashcard",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFlashcardDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/FlashcardResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "400 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "401 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "403 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "404 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "500 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Flashcards"
        ]
      },
      "delete": {
        "operationId": "FlashcardsController_deleteFlashcard",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "message": {
                              "type": "string",
                              "example": "Operation completed successfully"
                            }
                          },
                          "required": [
                            "message"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "401 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "403 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "404 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "500 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Flashcards"
        ]
      }
    },
    "/api/v1/flashcards/{id}/review": {
      "post": {
        "operationId": "FlashcardsController_reviewFlashcard",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/FlashcardResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "400 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "401 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "403 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "404 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "500 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Flashcards"
        ]
      }
    },
    "/api/v1/flashcards/{id}/mastered": {
      "post": {
        "operationId": "FlashcardsController_markAsMastered",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/FlashcardResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "401 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "403 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "404 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "500 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Flashcards"
        ]
      }
    },
    "/api/v1/sessions/flashcard/{flashcardId}": {
      "post": {
        "operationId": "SessionsController_createSession",
        "parameters": [
          {
            "name": "flashcardId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/CreateSessionResponseData"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "400 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "401 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "403 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "404 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "500 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Sessions"
        ]
      }
    },
...
  },
  "info": {
    "title": "FLENVN API",
    "description": "API documentation for the FLENVN backend services Includes inferred response schemas for frontend integration. Successful responses are wrapped by ResponseInterceptor as { success, data, timestamp }.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT access token"
      }
    },
    "schemas": {
...
      "CreateFlashcardDto": {
        "type": "object",
        "properties": {
          "word": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "resilient"
          },
          "partOfSpeech": {
            "type": "string",
            "enum": [
              "noun",
              "verb",
              "adjective",
              "adverb",
              "pronoun",
              "preposition",
              "conjunction",
              "interjection",
              "determiner"
            ],
            "example": "noun"
          },
          "pronunciation": {
            "type": "string",
            "example": "/rɪˈzɪl.i.ənt/"
          },
          "definition": {
            "type": "string",
            "example": "Able to recover quickly from difficulties."
          },
          "translation": {
            "type": "string",
            "example": "linh hoat"
          },
          "audioUrl": {
            "type": "string",
            "example": "https://cdn.example.com/audio.mp3"
          },
          "imageUrl": {
            "type": "string",
            "example": "https://cdn.example.com/image.png"
          },
          "example": {
            "type": "string",
            "example": "She remained resilient under pressure."
          },
          "exampleTranslation": {
            "type": "string",
            "example": "Co ay van kien cu duoi ap luc."
          },
          "bookId": {
            "type": "string",
            "example": "a9c40a35-7b10-4ec7-b86d-a8eeef1c3ad5"
          }
        },
        "required": [
          "word"
        ]
      },
      "UpdateFlashcardDto": {
        "type": "object",
        "properties": {
          "word": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "resilient"
          },
          "partOfSpeech": {
            "type": "string",
            "enum": [
              "noun",
              "verb",
              "adjective",
              "adverb",
              "pronoun",
              "preposition",
              "conjunction",
              "interjection",
              "determiner"
            ],
            "example": "noun"
          },
          "pronunciation": {
            "type": "string",
            "example": "/rɪˈzɪl.i.ənt/"
          },
          "definition": {
            "type": "string",
            "example": "Able to recover quickly from difficulties."
          },
          "translation": {
            "type": "string",
            "example": "linh hoat"
          },
          "audioUrl": {
            "type": "string",
            "example": "https://cdn.example.com/audio.mp3"
          },
          "imageUrl": {
            "type": "string",
            "example": "https://cdn.example.com/image.png"
          },
          "example": {
            "type": "string",
            "example": "She remained resilient under pressure."
          },
          "exampleTranslation": {
            "type": "string",
            "example": "Co ay van kien cu duoi ap luc."
          }
        }
      },
...
      "ApiEnvelope": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {},
          "message": {
            "type": "string",
            "example": "Optional message"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-01T10:00:00.000Z"
          }
        },
        "required": [
          "success",
          "data",
          "timestamp"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-01T10:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/auth/login"
          },
          "method": {
            "type": "string",
            "example": "POST"
          },
          "message": {
            "type": "string",
            "example": "Invalid credentials"
          },
          "details": {}
        },
        "required": [
          "statusCode",
          "timestamp",
          "path",
          "method",
          "message"
        ]
      },
...
      "FlashcardResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "word": {
            "type": "string",
            "example": "serendipity"
          },
          "partOfSpeech": {
            "type": "string",
            "nullable": true,
            "example": "noun"
          },
          "pronunciation": {
            "type": "string",
            "nullable": true,
            "example": "/ˌserənˈdipəti/"
          },
          "definition": {
            "type": "string",
            "nullable": true,
            "example": "The occurrence of events by chance in a happy way."
          },
          "translation": {
            "type": "string",
            "nullable": true,
            "example": "sự tình cờ may mắn"
          },
          "audioUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://example.com/audio.mp3"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://example.com/image.jpg"
          },
          "example": {
            "type": "string",
            "nullable": true,
            "example": "Finding that book was pure serendipity."
          },
          "exampleTranslation": {
            "type": "string",
            "nullable": true,
            "example": "Tìm thấy cuốn sách đó là một sự tình cờ may mắn."
          },
          "easeFactor": {
            "type": "number",
            "example": 2.5
          },
          "interval": {
            "type": "number",
            "example": 1
          },
          "repetitions": {
            "type": "number",
            "example": 0
          },
          "nextReviewDate": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-01T10:00:00.000Z",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "learning",
              "reviewing",
              "mastered"
            ],
            "example": "new"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "bookId": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-01T10:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-01T10:00:00.000Z"
          }
        }
      },
      "FlashcardStatsResponse": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 120
          },
          "new": {
            "type": "number",
            "example": 30
          },
          "learning": {
            "type": "number",
            "example": 20
          },
          "reviewing": {
            "type": "number",
            "example": 60
          },
          "mastered": {
            "type": "number",
            "example": 10
          },
          "dueForReview": {
            "type": "number",
            "example": 8
          }
        }
      },
...
      "CreateSessionResponseData": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Session recorded successfully"
          },
          "session": {
            "$ref": "#/components/schemas/SessionResponse"
          }
        }
      },
...
    }
  }
}